+++ src/BinaryFile.h
@@ -30,9 +30,9 @@ namespace tivars
             close();
         }
 
-        uchar get_raw_byte();
-        data_t get_raw_bytes(uint bytes);
-        std::string get_string_bytes(uint bytes);
+        uint8_t get_raw_byte();
+        data_t get_raw_bytes(size_t bytes);
+        std::string get_string_bytes(size_t bytes);
         size_t size() const;
         void close();
